==> Oklahoma/Overlays/0_OKRegulatorySourceDataPreprocess.ipynb <==
# OK Regulatory Data

Preprocessing input data for a smoother upload experience of the state data to the WaDE 2.0 database.
Using geopandas to read in shp file, and coverting to WKT for ReportingUnit geometry.

Notes:
- WKT will be in it's own separate geometery dataframe.

#Needed Libararies
import os
import numpy as np
import pandas as pd
from datetime import datetime
import geopandas as gpd # the library that lets us read in shapefiles
pd.set_option('display.max_columns', 999)  # How to display all columns of a Pandas DataFrame in Jupyter Notebook

#Setting work directory, reading inputs, creating dataframe
workingDir = "C:/Users/rjame/Documents/WSWC Documents/MappingStatesDataToWaDE2.0/Oklahoma/Regulatory/RawInputData"
os.chdir(workingDir)

#CSV input file
fileInput = "Surface Water Standards Details_input.csv"

df = pd.read_csv(fileInput)
print(len(df))
df.head()

df['in_OversightAgency'] = "WQ Management Segmen " + df['WQMSEG'].astype(str)
df.head()

df_out = df

# Shapefile Data

# Interstate Stream Compact Regions

ShapeFile = "ShapefileData/Surface_Water_Standards.shp"
dfshp = gpd.read_file(ShapeFile)

print(len(dfshp))
dfshp.head()

# Interstate Stream Compact Regions output shapefile dataframe
columnList = ['OBJECTID', 
              'NAME',
              'geometry']
dfshp_out = pd.DataFrame(columns=columnList, index=dfshp.index)

dfshp_out['OBJECTID'] = dfshp['OBJECTID']
dfshp_out['NAME'] = dfshp['NAME'].str.title()
dfshp_out['geometry'] = dfshp['geometry']

print(len(dfshp_out))
dfshp_out.head(3)

### Inspect Output Data & Export

with pd.option_context('display.max_rows', None, 'display.max_columns', None):
    print(df_out.dtypes)

with pd.option_context('display.max_rows', None, 'display.max_columns', None):
    print(dfshp_out.dtypes)

# Export out to CSV.
df_out.to_csv('P_okRegMaster.csv', index=False) # The output.
dfshp_out.to_csv('P_okRegGeometry.csv', index=False) # The output geometry.
